Add URL shortening support (#3)
Added URL shortening support with randomly generated slugs. Each
Shortened URL receives a unique slug which is used to reference the URL
in the future. URLs are validated and must comply with the
requirements before they are shortened.

When accessing the shortened url `/:slug` path, users are redirected to
the real URL.

UI is built using [Bulma CSS framework](https://bulma.io/).